Initial program to retrieve gmail user inbox data
I retrieved the data of a users' inbox by utilizing the scope that was provided
by google along with the username and password given to me from the user.  I
then opened a connection to the url and passed in the username and password
encoded in Base64.  Once the connection was open, I used a BufferedReader to
store the stream into a variable "xml" which is outputed to the console.